# src/lib/openapi.yaml
openapi: 3.0.0
info:
  title: PetStayTion API
  version: 1.0.0
  description: API for pet care service management
servers:
  - url: /
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/reviews:
    post:
      summary: Create a review for a service provider
      description: |
        Allows clients to create a review (rating 1-5 and optional comment) for a service provider after a booking with PAID status.
        - Only clients can create reviews.
        - Booking must have PAID status.
        - Each booking can only be reviewed once.
        - Rating must be between 1 and 5.
        - Requires valid Firebase authentication token.
      tags: [Reviews]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bookingId
                - rating
              properties:
                bookingId:
                  type: integer
                  description: ID of the PAID booking to review
                  example: 123
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  description: Rating from 1 to 5
                  example: 5
                comment:
                  type: string
                  maxLength: 1000
                  description: Optional comment about the service
                  example: "Excellent service! Very professional and caring."
      responses:
        "200":
          description: Review created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Review created successfully"
                  review:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      rating:
                        type: integer
                        example: 5
                      comment:
                        type: string
                        nullable: true
                        example: "Excellent service! Very professional and caring."
                      createdAt:
                        type: string
                        format: date-time
                        example: "2025-11-18T12:00:00Z"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "You can only review bookings with PAID status"
        "401":
          description: Unauthorized (invalid or missing token)
        "403":
          description: Forbidden (user is not a client)
        "404":
          description: Booking not found or no permission
        "500":
          description: Internal server error
  /api/reviews/{id}:
    delete:
      summary: Dead that review you dropped, nigga
      description: You the one who wrote it, so you the only nigga that can wipe it. Anybody else move funny, they get lined out the whole app.
      tags: [Reviews]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Review ID you finna smoke
      responses:
        "200":
          description: Gone. That shit deleted, nigga.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "Review took a dirt nap"
        "401":
          description: You ain't logged in, who the fuck is you nigga?
        "403":
          description: That ain't yo review nigga, touch it and you catch a permanent ban
        "404":
          description: Shit already ghost, can't find it
        "500":
          description: Server down bad, whole block bleedin' rn
    put:
      summary: Switch up on yo old take, nigga
      description: Mind changed? Flip the rating, rewrite the comment, do what you gotta do. But try to touch another nigga shit and we got real problems.
      tags: [Reviews]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The review ID you own, nigga
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  description: How many stars you givin' it now, nigga
                  example: 1
                comment:
                  type: string
                  nullable: true
                  maxLength: 1000
                  description: New shit you spittin' or null to clear it
                  example: "On some real nigga shit, this trash. I was trippin' when I said it was fire."
              anyOf:
                - required: [rating]
                - required: [comment]
      responses:
        "200":
          description: Updated clean, nigga. It's official.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "Review flipped, no 🧢"
                  review:
                    type: object
                    properties:
                      id:
                        type: integer
                      rating:
                        type: integer
                      comment:
                        type: string
                        nullable: true
                      createdAt:
                        type: string
                        format: date-time
        "400":
          description: Request lookin' dumb, fix that shit nigga
        "401":
          description: Ain't signed in, get yo credentials right
        "403":
          description: Tryna edit another nigga review? You wild for that
        "404":
          description: That review already in the grave
        "500":
          description: Backend crashin' out, trap house on fire rn
...
  /api/user/attributes:
    post:
      summary: Register a new user profile
      description: |
        Creates a new user in the database with attributes such as name, surname, phone number, and city.
        A client record is created automatically for the user.
        Authentication and password management are handled by Firebase — this endpoint only stores the user's Firebase UID and profile data.
      tags: [User]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firebaseUid
                - email
                - cityId
              properties:
                firebaseUid:
                  type: string
                  description: Firebase UID of the user (from Firebase Auth)
                  example: "abcd1234efgh5678"
                email:
                  type: string
                  format: email
                  description: User email (must be unique)
                  example: "jane.doe@example.com"
                firstName:
                  type: string
                  description: Optional first name
                  example: "Jane"
                lastName:
                  type: string
                  description: Optional last name
                  example: "Doe"
                phoneNumber:
                  type: string
                  description: Optional 9-digit phone number
                  example: "123456789"
                cityId:
                  type: integer
                  description: Required ID of the city (foreign key)
                  example: 1
      responses:
        "200":
          description: User profile created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 42
                      firebaseUid:
                        type: string
                        example: "abcd1234efgh5678"
                      firstName:
                        type: string
                        nullable: true
                        example: "Jane"
                      lastName:
                        type: string
                        nullable: true
                        example: "Doe"
                      email:
                        type: string
                        example: "jane.doe@example.com"
                      phoneNumber:
                        type: string
                        nullable: true
                        example: "123456789"
                      city:
                        type: object
                        properties:
                          idCity:
                            type: integer
                            example: 1
                          name:
                            type: string
                            example: "Warszawa"
                          imageUrl:
                            type: string
                            nullable: true
                            example: "https://example.com/city.jpg"
        "400":
          description: Invalid input (missing required fields, invalid phone number, or city not found)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Firebase UID, email, and cityId are required"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "User already exists"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
...